MaterialUI/backend/requirements.txt
fastapi
uvicorn
boto3
flask
python-dotenv
faiss-cpu
transformers
torch
google-generativeai
